+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Create new setting for title level requests. Refs UICIRC-698.
 * UI tests replacement with RTL/Jest for component `NoticesList` (EditSections). Refs UICIRC-641.
 * Locations display as `{code} {name}` and thus should be sorted by code, name. Refs UICIRC-703.
+* Add RTL/Jest testing for `PatronNoticeDetail` component in `settings/PatronNotices`. Refs UICIRC-643.
 
 ## [6.0.0] (https://github.com/folio-org/ui-circulation/tree/v6.0.0) (2021-09-30)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v5.1.1...v6.0.0)